MP P1 port (step 4f): fix mac's 4 clang-strictness divergences in the vp* subset
Caught by mac's clang validation (gcc -fpermissive only warned):
- vplist.h: drop C++17-illegal 'register' (clang hard-errors; hit all 6 TUs)
- vpengine.cpp:450 ptr-cast-to-DWORD under %s -> pass the string (LP64 truncated the char*);
  :849/:1667 (DWORD)link -> (DWORD)(uintptr_t)link for %08lx (LP64-safe)
- vdmplay.cpp: Win16/Win32s Universal-Thunk block (vpUtStartup + CUtNotifyQueue::PostMessage,
  FARPROC casts, missing rettype) #ifndef WIN32 -> #ifdef _WIN16 (excluded on Win32 AND POSIX;
  vput16.cpp the only user, not in POSIX build)
All 7 TUs (6 + vpnet) still gcc -fsyntax-only clean. Windows untouched.

+++ tools/vdmplay/vpengine.cpp
@@ -447,7 +447,7 @@ BOOL CVpSession::KnockOutPlayer( VPPLAYERID id, plrInfoMsg* msg, CRemoteWS* ws )
 
 BOOL CVpSession::KillPlayer( VPPLAYERID id ) {
     LogV( m_log, "CVpSession::KillPlayer(%s) %u",
-          (DWORD)( IsLocal() ? "Local" : "Remote" ), id );
+          ( IsLocal() ? "Local" : "Remote" ), id );   // %s wants char* — dropped (DWORD) cast (LP64-truncated the ptr)
 
     if ( !IsLocal() ) {
         SetError( VP_ERR_REMOTE_SESSION );
@@ -846,7 +846,7 @@ BOOL SimulateLeave( CPlayer* p, LPVOID ctx ) {
 
 void CLocalSession::OnDisconnect( CNetLink* link ) {
     LogV( m_log, "ClocalSession::OnDisconnect for link %08lx\n",
-          (DWORD)link );
+          (DWORD)(uintptr_t)link );   // LP64: cast via uintptr_t (ptr->DWORD direct is a clang error)
 
     if ( m_broadcastLink == link ) {
         HandleNetDown();
@@ -1664,7 +1664,7 @@ void CRemoteSession::OnAccept( CNetLink* link ) {
 
 void CRemoteSession::OnDisconnect( CNetLink* link ) {
     VPENTER( CRemoteSession::OnDisconnect );
-    LogV( m_log, "RemoteSession::OnDisconnect link %08lx", (DWORD)link );
+    LogV( m_log, "RemoteSession::OnDisconnect link %08lx", (DWORD)(uintptr_t)link );   // LP64-safe cast
 
     if ( m_broadcastLink == link ) {
         m_connected = FALSE;
